refactor premium gas service
